% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_pca.R
\name{comp_power_pca}
\alias{comp_power_pca}
\title{Power Transform Principal Component Analysis for PSA Comparison}
\usage{
comp_power_pca(X, alpha = 1/2)
}
\arguments{
\item{X}{a data matrix.}

\item{alpha}{a scalar \eqn{alpha>0} for power transform.}
}
\value{
A list
\itemize{
\item pts.approx a list of lower dimensional representations with respect to the original basis
\item scores a matrix of scores.
\item rss a vector of residual sums of squares.
\item modes a list of modes of variation. The \eqn{r}th element of the list is the difference of
rank \eqn{r} approximations to rank \eqn{r-1} approximations.
\item loadings a matrix of loading vectors.
\item center mean of the data
}
}
\description{
A wrapper function of \code{princomp()} applied to power transformed data.
}
\seealso{
\code{\link[=comp_pca]{comp_pca()}}, \code{\link[=psa]{psa()}}
}
