% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{gridsearch_pair}
\alias{gridsearch_pair}
\title{Find the best weight for merging two vertices by grid search}
\usage{
gridsearch_pair(
  X,
  v1,
  v2,
  testweights = seq(0, 1, length.out = 100),
  evaluator,
  ...
)
}
\arguments{
\item{X}{matrix of data points for a simplex with vertices at the basis coordinates.}

\item{v1}{an index of a vertex (a basis vector)}

\item{v2}{an index of a vertex (a basis vector)}

\item{testweights}{A vector of weights to try.}

\item{evaluator}{A function that evaluates each proposed merge and return a single non-negative value for the quality of the merge. Arguments of \code{X}, \code{v1}, \code{v2}, and \code{w} at least}

\item{...}{Passed to \code{evaluator}}
}
\description{
Find the best weight for merging two vertices by grid search
}
