% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectsubsimplex.R
\name{projectsubsimplex}
\alias{projectsubsimplex}
\title{Project to a Subsimplex}
\usage{
projectsubsimplex(X, V = diag(ncol(X)), v1, v2, w)
}
\arguments{
\item{X}{a data matrix whose coordinates are computed with respect to \code{V}.}

\item{V}{matrix for vertices of the reduced simplex.}

\item{v1, v2}{vertices to be merged.}

\item{w}{weight for merge.}
}
\description{
Project to a Subsimplex
}
\details{
Following Version 4 of the descriptions.
The result of the projection are points with mass (i.e. value) in the newly created vertex equal to the sum of the mass (i.e. value) in the components specified by \code{v1} and \code{v2}.
The new vertex is computed as \code{w * V[v1, ] + (1-w) * V[v2, ]}, which is the convex combination of two vertices and is \emph{not} a vector with L2-norm equal to 1.
Scores a signed as positive if the difference between the input point and the projected point has a positive inner-product with the merge direction \code{V[v2, ] - V[v1, ]}.
}
