% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psoa.R
\name{psoa}
\alias{psoa}
\title{Principle Sub Orthant Analysis for the Simplex}
\usage{
psoa(
  X,
  V = diag(ncol(X)),
  testweights = seq(0, 1, length.out = 100),
  merges = NULL,
  maxsteps = min(ncol(X), nrow(merges)) - 1
)
}
\arguments{
\item{X}{matrix of data points for a simplex with vertices at the basis coordinates. Each row an observation.}

\item{V}{matrix of vertices in the basis of the \emph{original} data. Each row a vertex corresponding to the \emph{columns} of \code{X}.}

\item{testweights}{A vector of weights to try.}

\item{merges}{if non-NULL, then a data frame of prespecified merges.}

\item{maxsteps}{maximum number of merges to compute.}
}
\description{
Principle Sub Orthant Analysis for the Simplex
}
