% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_simplex.R
\name{to_simplex}
\alias{to_simplex}
\title{Projection onto the Simplex}
\usage{
to_simplex(X)
}
\arguments{
\item{X}{a data matrix.}
}
\value{
a matrix of the same dimension as \code{X}.
}
\description{
Projection onto the Simplex
}
